ai_gateway_backend_extracted/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
cryptography
httpx